requirements.txt
spotipy
requests
Pillow
dotenv
